# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mari Servis API",
        "contact": {
            "name": "RODERICK HALIM",
            "url": "https://uph.edu"
        },
        "version": "1.0"
    },
    "host": "mariservis.dev",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Register User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "phoneNumber": {
                                    "type": "string"
                                },
                                "path": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User has been log in successfully"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        }
    },
...
}